Toradex09/T09V2/requirements.txt
python-can>=4.0.0
paho-mqtt
